requirements.txt
Flask
mysql-connector-python
scikit-learn
pandas
Flask-Cors
python-dotenv